Remove unnecessary variant creation calls
Clingo’s variants provide a static member function “make” to create new
variants. As it turns out, calling this function explicitly isn’t
necessary because the variants provide constructors taking an rvalue of
any of the contained types. In other words, values of the allowed types
can be assigned to variants directly, and they can be directly returned
where a suitable variant is expected.