Fix unreachable code handling
While this line of code should be unreachable, it’s not acceptable to
just return any value. In case there actually is a bug somewhere and
this line of code is reached for some reason, it will be very hard to
debug what’s actually going on. It’s a better choice to throw an
exception to indicate that this outcome is not expected.